# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Apache Livy successor working on Spark 4.0 and Spark Connect.",
        "title": "Livy-Next API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8998",
    "basePath": "/",
    "paths": {
        "/sessions": {
            "get": {
                "description": "Get a list of all active interactive sessions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "List all active sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SessionsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new interactive session and connect to Spark Connect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Create a new session",
                "parameters": [
                    {
                        "description": "Create Session Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/session.Session"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to connect to Spark Connect",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sessions/{id}": {
            "get": {
                "description": "Get details and state of a specific session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Get session details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/session.Session"
                        }
                    },
                    "400": {
                        "description": "Invalid session ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Close and terminate a specific session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Delete session",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session deleted msg",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid session ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sessions/{id}/statements": {
            "get": {
                "description": "List all statements submitted to a session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statements"
                ],
                "summary": "List statements",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StatementsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid session ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Submit a SQL statement for execution in a session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statements"
                ],
                "summary": "Submit a statement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Submit Statement Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateStatementRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/session.Statement"
                        }
                    },
                    "400": {
                        "description": "Invalid payload/ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Session not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sessions/{id}/statements/{statementId}": {
            "get": {
                "description": "Get execution state and results of a statement",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statements"
                ],
                "summary": "Get statement details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Statement ID",
                        "name": "statementId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/session.Statement"
                        }
                    },
                    "400": {
                        "description": "Invalid session/statement ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Session/Statement not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/sessions/{id}/statements/{statementId}/cancel": {
            "post": {
                "description": "Cancel a statement execution inside a session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statements"
                ],
                "summary": "Cancel a statement",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Statement ID",
                        "name": "statementId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg: cancelled",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid session/statement ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Session/Statement not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CreateSessionRequest": {
            "type": "object",
            "properties": {
                "conf": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "jars": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "proxyUser": {
                    "type": "string"
                }
            }
        },
        "api.CreateStatementRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "api.SessionsResponse": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "integer"
                },
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/session.Session"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.StatementsResponse": {
            "type": "object",
            "properties": {
                "statements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/session.Statement"
                    }
                },
                "total_statements": {
                    "type": "integer"
                }
            }
        },
        "session.Session": {
            "type": "object",
            "properties": {
                "appId": {
                    "type": "string"
                },
                "appInfo": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "description": "\"spark\", \"pyspark\", \"sparkr\"",
                    "type": "string"
                },
                "lastActivity": {
                    "type": "string"
                },
                "log": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "proxyUser": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/session.SessionState"
                },
                "statements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/session.Statement"
                    }
                }
            }
        },
        "session.SessionState": {
            "type": "string",
            "enum": [
                "starting",
                "idle",
                "busy",
                "dead",
                "shutting_down"
            ],
            "x-enum-varnames": [
                "SessionStarting",
                "SessionIdle",
                "SessionBusy",
                "SessionDead",
                "SessionShuttingDown"
            ]
        },
        "session.Statement": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "completed": {
                    "description": "Milliseconds epoch",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "output": {
                    "$ref": "#/definitions/session.StatementOutput"
                },
                "progress": {
                    "type": "number"
                },
                "started": {
                    "description": "Milliseconds epoch",
                    "type": "integer"
                },
                "state": {
                    "$ref": "#/definitions/session.StatementState"
                }
            }
        },
        "session.StatementOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "e.g. \"application/json\" -\u003e QueryResult or \"text/plain\" -\u003e string",
                    "type": "object",
                    "additionalProperties": true
                },
                "ename": {
                    "type": "string"
                },
                "evalue": {
                    "type": "string"
                },
                "execution_count": {
                    "type": "integer"
                },
                "status": {
                    "description": "\"ok\" or \"error\"",
                    "type": "string"
                },
                "traceback": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "session.StatementState": {
            "type": "string",
            "enum": [
                "waiting",
                "running",
                "available",
                "error",
                "cancelling",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatementWaiting",
                "StatementRunning",
                "StatementAvailable",
                "StatementError",
                "StatementCancelling",
                "StatementCancelled"
            ]
        }
    }
}